LPS-165763 Rename
