+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/exporter/LayoutsExporterImpl.java
@@ -28,11 +28,11 @@ import com.fasterxml.jackson.databind.util.ISO8601DateFormat;
 
 import com.liferay.headless.delivery.dto.v1_0.PageDefinition;
 import com.liferay.layout.exporter.LayoutsExporter;
-import com.liferay.layout.internal.headless.delivery.dto.v1_0.converter.util.DisplayPageTemplateUtil;
-import com.liferay.layout.internal.headless.delivery.dto.v1_0.converter.util.MasterPageUtil;
-import com.liferay.layout.internal.headless.delivery.dto.v1_0.converter.util.PageTemplateCollectionUtil;
-import com.liferay.layout.internal.headless.delivery.dto.v1_0.converter.util.PageTemplateUtil;
-import com.liferay.layout.internal.headless.delivery.dto.v1_0.converter.util.UtilityPageTemplateUtil;
+import com.liferay.layout.internal.headless.delivery.dto.v1_0.util.DisplayPageTemplateUtil;
+import com.liferay.layout.internal.headless.delivery.dto.v1_0.util.MasterPageUtil;
+import com.liferay.layout.internal.headless.delivery.dto.v1_0.util.PageTemplateCollectionUtil;
+import com.liferay.layout.internal.headless.delivery.dto.v1_0.util.PageTemplateUtil;
+import com.liferay.layout.internal.headless.delivery.dto.v1_0.util.UtilityPageTemplateUtil;
 import com.liferay.layout.page.template.constants.LayoutPageTemplateEntryTypeConstants;
 import com.liferay.layout.page.template.constants.LayoutPageTemplateExportImportConstants;
 import com.liferay.layout.page.template.model.LayoutPageTemplateCollection;
