+++ modules/apps/layout/layout-impl/src/main/java/com/liferay/layout/internal/headless/delivery/dto/v1_0/util/UtilityPageTemplateTypeUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.internal.headless.delivery.dto.v1_0.converter.util;
+package com.liferay.layout.internal.headless.delivery.dto.v1_0.util;
 
 import com.liferay.layout.utility.page.constants.LayoutUtilityPageEntryConstants;
 import com.liferay.portal.kernel.util.HashMapBuilder;
